+++ react/javascript/src/index.ts
@@ -1,26 +1,24 @@
-import GherkinDocumentList from './components/app/GherkinDocumentList'
-import FilteredResults from './components/app/FilteredResults'
-import QueriesWrapper from './components/app/QueriesWrapper'
-import StatusIcon from './components/gherkin/StatusIcon'
 import filterByStatus from './filter/filterByStatus'
 import GherkinQueryContext from './GherkinQueryContext'
 import CucumberQueryContext from './CucumberQueryContext'
-import SearchQueryContext from './SearchQueryContext'
+import SearchQueryContext, {
+  WindowUrlApi,
+  SearchQueryUpdateFn,
+  SearchQueryProps,
+  searchFromURLParams,
+} from './SearchQueryContext'
 import EnvelopesQueryContext, { EnvelopesQuery } from './EnvelopesQueryContext'
-import MDG from './components/gherkin/MDG'
-import GherkinDocument from './components/gherkin/GherkinDocument'
 
+export * as components from './components'
 export {
-  GherkinDocumentList,
-  QueriesWrapper,
   GherkinQueryContext,
   CucumberQueryContext,
-  SearchQueryContext,
-  StatusIcon,
   EnvelopesQueryContext,
   EnvelopesQuery,
-  FilteredResults,
   filterByStatus,
-  MDG,
-  GherkinDocument,
+  SearchQueryContext,
+  WindowUrlApi,
+  SearchQueryUpdateFn,
+  SearchQueryProps,
+  searchFromURLParams,
 }