+++ react/javascript/src/stories/GherkinDocument.stories.tsx
@@ -4,10 +4,7 @@ import { Meta, Story } from '@storybook/react'
 import * as messages from '@cucumber/messages'
 import { Query as CucumberQuery } from '@cucumber/query'
 import { Query as GherkinQuery } from '@cucumber/gherkin-utils'
-import GherkinDocumentList from '../components/app/GherkinDocumentList'
-import QueriesWrapper from '../components/app/QueriesWrapper'
-import { EnvelopesQuery } from '../../src'
-import GherkinDocument from '../components/gherkin/GherkinDocument'
+import { EnvelopesQuery, components } from '../../src'
 
 import '../styles/styles.scss'
 
@@ -18,13 +15,19 @@ import hooks from '../../acceptance/hooks/hooks.feature'
 import markdown from '../../acceptance/markdown/markdown.feature.md'
 import minimal from '../../acceptance/minimal/minimal.feature'
 import parameterTypes from '../../acceptance/parameter-types/parameter-types.feature'
+import pending from '../../acceptance/pending/pending.feature'
+import retry from '../../acceptance/retry/retry.feature'
 import rules from '../../acceptance/rules/rules.feature'
+import skipped from '../../acceptance/skipped/skipped.feature'
 import stacktTraces from '../../acceptance/stack-traces/stack-traces.feature'
+import undefinedEnvelopes from '../../acceptance/undefined/undefined.feature'
 import unknownParameterTypes from '../../acceptance/unknown-parameter-type/unknown-parameter-type.feature'
+
+const { QueriesWrapper, GherkinDocumentList } = components.app
 
 export default {
   title: 'GherkinDocument',
-  component: GherkinDocument,
+  component: components.gherkin.GherkinDocument,
 } as Meta
 
 type Props = {
@@ -80,14 +83,34 @@ ParameterTypes.args = {
   envelopes: parameterTypes,
 }
 
+export const Pending = Template.bind({})
+Pending.args = {
+  envelopes: pending,
+}
+
+export const Retry = Template.bind({})
+Retry.args = {
+  envelopes: retry,
+}
+
 export const Rules = Template.bind({})
 Rules.args = {
   envelopes: rules,
 }
 
+export const Skipped = Template.bind({})
+Skipped.args = {
+  envelopes: skipped,
+}
+
 export const StackTraces = Template.bind({})
 StackTraces.args = {
   envelopes: stacktTraces,
+}
+
+export const Undefined = Template.bind({})
+Undefined.args = {
+  envelopes: undefinedEnvelopes,
 }
 
 export const UnknownParameterTypes = Template.bind({})