+++ scripts/functions.sh
@@ -138,12 +138,12 @@ function update_go_library_version()
     pushd ${root_dir}
     git ls-files "${root_dir}/**/go.mod" | while read go_mod; do
       # Edit require directives
-      sed -Ei "s/github\.com\/cucumber\/${libname}-go(\/v[[:digit:]]+)? v.*/github\.com\/cucumber\/${libname}-go\/v${major} v${version}/" ${go_mod}
+      sed -Ei "s/github\.com\/cucumber\/common\/${libname}\/go(\/v[[:digit:]]+)? v.*/github\.com\/cucumber\/common\/${libname}\/go\/v${major} v${version}/" ${go_mod}
       # Edit replace directives
-      sed -Ei "s/github\.com\/cucumber\/${libname}-go(\/v[[:digit:]]+)? =>/github\.com\/cucumber\/${libname}-go\/v${major} =>/" ${go_mod}
+      sed -Ei "s/github\.com\/cucumber\/common\/${libname}\/go(\/v[[:digit:]]+)? =>/github\.com\/cucumber\/common\/${libname}\/go\/v${major} =>/" ${go_mod}
     done
     git ls-files "${root_dir}/**/*.go" | while read go_mod; do
-      sed -Ei "s/github\.com\/cucumber\/${libname}-go(\/v[[:digit:]]+)?/github\.com\/cucumber\/${libname}-go\/v${major}/" ${go_mod}
+      sed -Ei "s/github\.com\/cucumber\/common\/${libname}\/go(\/v[[:digit:]]+)?/github\.com\/cucumber\/common\/${libname}\/go\/v${major}/" ${go_mod}
       # sed -i "s/github.com\/cucumber\/${libname}-go\/v[[:digit:]]\+/github.com\/cucumber\/${libname}-go\/v${major}/" ${go_mod}
     done
     popd