+++ suggest/javascript/src/index.ts
@@ -1,0 +1,7 @@
+import buildStepDocuments from './buildStepDocuments'
+import lspCompletionSnippet from './lspCompletionSnippet'
+import { Index, StepDocument, StepSegment } from './types'
+
+export { buildStepDocuments, lspCompletionSnippet }
+
+export type { Index, StepDocument, StepSegment }